refactor(code-examples): make it a gatsby theme (#419)
* refactor(code-examples): move all code examples related stuff into transformer package

* refactor: make code examples a gatsby theme

* docs(code-examples): remove code examples docs from base theme, improve docs
