+++ packages/gatsby-theme-code-examples/gatsby-config.js
@@ -1,0 +1,13 @@
+const path = require('path');
+
+module.exports = {
+  plugins: [
+    {
+      resolve: `gatsby-source-filesystem`,
+      options: {
+        name: `codeExamples`,
+        path: path.resolve('./src/code-examples'),
+      },
+    },
+  ],
+};
