+++ packages/gatsby-theme-code-examples/index.js
