+++ packages/gatsby-theme-code-examples/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@commercetools-docs/gatsby-theme-code-examples",
+  "version": "2.5.0",
+  "license": "MIT",
+  "private": false,
+  "publishConfig": {
+    "access": "public"
+  },
+  "description": "Gatsby theme for commercetools documentation providing the functionalities to render Code Examples in MDX",
+  "homepage": "https://github.com/commercetools/commercetools-docs-kit/tree/master/packages/gatsby-theme-code-examples#readme",
+  "keywords": [
+    "gatsby",
+    "gatsby-plugin",
+    "gatsby-theme",
+    "code",
+    "examples"
+  ],
+  "main": "src/index.js",
+  "dependencies": {
+    "@commercetools-docs/ui-kit": "2.5.0",
+    "prop-types": "15.7.2"
+  },
+  "devDependencies": {
+    "gatsby": "2.21.11",
+    "react": "16.13.1",
+    "react-dom": "16.13.1"
+  },
+  "peerDependencies": {
+    "@commercetools-docs/gatsby-theme-docs": "2.x",
+    "gatsby": "2.x",
+    "react": "16.x",
+    "react-dom": "16.x"
+  }
+}
