+++ packages/gatsby-theme-code-examples/src/@commercetools-docs/gatsby-theme-docs/overrides/markdown-components.js
@@ -1,0 +1,4 @@
+import CodeExample from '../../../code-example';
+import MultiCodeExample from '../../../multi-code-example';
+
+export default { CodeExample, MultiCodeExample };
