+++ packages/gatsby-theme-code-examples/src/use-code-examples.js
@@ -1,0 +1,17 @@
+import { useStaticQuery, graphql } from 'gatsby';
+
+export default () => {
+  const queryResult = useStaticQuery(graphql`
+    query GetAllCodeExamplesQuery {
+      allCodeExample {
+        nodes {
+          name
+          language
+          path
+          content
+        }
+      }
+    }
+  `);
+  return queryResult.allCodeExample.nodes;
+};
