+++ packages/gatsby-theme-docs/src/components/index.js
@@ -1,12 +1,10 @@
 export { default as Anchor } from './anchor';
 export { default as BetaFlag } from './beta-flag';
 export { default as BurgerIcon } from './burger-icon';
-export { default as CodeExample } from './code-example';
 export { default as ContentPagination } from './content-pagination';
 export { default as GlobalNavigationLink } from './global-navigation-link';
 export { default as ErrorBoundary } from './error-boundary';
 export { default as Link, ExternalSiteLink } from './link';
-export { default as MultiCodeExample } from './multi-code-example';
 export { default as Overlay } from './overlay';
 export { default as ReleaseNotesSubscribeLinks } from './release-notes-subscribe-links';
 export { default as SEO } from './seo';
