--- packages/gatsby-theme-docs/src/hooks/use-code-examples.js
@@ -1,20 +1,0 @@
-import { useStaticQuery, graphql } from 'gatsby';
-
-export default () => {
-  const queryResult = useStaticQuery(
-    graphql`
-      query GetAllCodeExamplesQuery {
-        allCodeExample {
-          nodes {
-            name
-            language
-            path
-            content
-          }
-        }
-      }
-    `
-  );
-
-  return queryResult.allCodeExample.nodes;
-};
