--- packages/gatsby-transformer-code-examples/gatsby-config.js
@@ -1,15 +1,0 @@
-const path = require('path');
-
-module.exports = () => {
-  return {
-    plugins: [
-      {
-        resolve: 'gatsby-source-filesystem',
-        options: {
-          name: 'codeExamples',
-          path: path.resolve(`./src/code-examples`),
-        },
-      },
-    ],
-  };
-};
