--- packages/gatsby-transformer-code-examples/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@commercetools-docs/gatsby-transformer-code-examples",
-  "version": "2.5.0",
-  "license": "MIT",
-  "private": false,
-  "publishConfig": {
-    "access": "public"
-  },
-  "description": "Provides contents of nodes with source code mime types as CodeExample GraphQl types.",
-  "homepage": "https://github.com/commercetools/commercetools-docs-kit/tree/master/packages/gatsby-transformer-code-examples#readme",
-  "keywords": [
-    "gatsby",
-    "gatsby-plugin",
-    "gatsby-transformer",
-    "code",
-    "examples"
-  ]
-}
