+++ websites/docs-smoke-test/gatsby-config.js
@@ -17,5 +17,6 @@ module.exports = {
         additionalPrismLanguages: ['scala', 'csharp'],
       },
     },
+    '@commercetools-docs/gatsby-theme-code-examples',
   ],
 };
