+++ websites/docs-smoke-test/package.json
@@ -15,6 +15,7 @@
   },
   "dependencies": {
     "@commercetools-docs/gatsby-theme-docs": "2.5.0",
+    "@commercetools-docs/gatsby-theme-code-examples": "2.5.0",
     "gatsby": "2.21.11",
     "gatsby-cli": "2.12.8"
   }
